requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.24.0
scikit-image>=0.20.0

# Data handling and visualization
pandas>=1.5.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Gradio UI
gradio>=4.0.0

# Utilities
tqdm>=4.65.0
ipython>=8.12.0
jupyter>=1.0.0

# For medical imaging (optional but recommended)
pydicom>=2.3.0
SimpleITK>=2.2.0

# For model optimization
bitsandbytes>=0.39.0
peft>=0.3.0

# For API deployment (optional)
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6

# Additional utilities
scipy>=1.10.0
requests>=2.28.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
mock>=4.0.0